backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.81.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
